chore: rename cammel case
