+++ apps/server/src/ws.ts
@@ -8,9 +8,9 @@ import { createCoreInstance } from '@tg-search/core'
 import { createRouter, defineEventHandler, defineWebSocketHandler, getQuery } from 'h3'
 
 import { handleConnectionEvent, registerConnectionEventHandler } from './event-handlers/connection'
-import { handleDialogsEvent, registerDialogsEventHandler } from './event-handlers/dialogs'
+import { handleDialogsEvent, registerDialogsEventHandler } from './event-handlers/dialog'
 import { handleEntityEvent, registerEntityEventHandler } from './event-handlers/entity'
-import { handleMessageEvent, registerMessageEventHandler } from './event-handlers/messages'
+import { handleMessageEvent, registerMessageEventHandler } from './event-handlers/message'
 import { handleTakeoutEvent, registerTakeoutEventHandler } from './event-handlers/takeout'
 import { registerWsMessageRoute, routeWsMessage } from './routes'
 import { createResponse } from './utils/response'
